Trait wtf8_rs::wtf8::Wtf8Index

pub unsafe trait Wtf8Index: Debug + Clone + Sealed {
    fn get(self, slice: &Wtf8) -> Option<&Wtf8>;

    unsafe fn get_unchecked(self, slice: *const Wtf8) -> *const Wtf8;
}

A helper trait to do get operation on Wtf8.

This trait is sealed and not meant to be implemented by an user of this library.

Required methods

fn get(self, slice: &Wtf8) -> Option<&Wtf8>

unsafe fn get_unchecked(self, slice: *const Wtf8) -> *const Wtf8

Loading content...